Updating

+++ MCAL/5-ADC/ADC_interface.h
@@ -25,7 +25,7 @@ typedef struct
 
 void ADC_Init(void);
 uint16 ADC_u16StartConversionSynch(uint8 Copy_u8Channel, uint16* Copy_pu16Result);
-uint8 ADC_u16StartSingleConversionAsynch(uint8 Copy_u8Channel,uint16* Copy_pu16Result, void(*Copy_pvNotificationFunc)(void));
+uint16 ADC_u16StartSingleConversionAsynch(uint8 Copy_u8Channel,uint16* Copy_pu16Result, void(*Copy_pvNotificationFunc)(void));
 uint8 ADC_u8StartChainConversionAsynch(ADC_ChainConv_t* Copy_pstChain);
 
 #endif /* ADC_INTERFACE_H_ */
